Reducde max UDP/NFS buffer size as it sometimes crashes with larger buffer.
Its a very very strange crash and I've spent quite a while trying to fix it, but its in the
supplied network code and horrible stuff. It really seems to be a bug in the supplied code
which is funny since I think the network code is actually provided by Intel.